Fixes Beaker Breaking (#7676)
Due to some of the, er, interesting decisions made when writing the code for beakers shattering when you throw them, not only was it possible to shatter containers made of non-glass materials into glass fragments - such as cryostasis and bluespace beakers, and wooden buckets - but most of the actual glass beakers were impossible to break with your bare hands.

I have corrected both these issues in addition to making it possible for glass beakers that are tougher to still shatter even if you don't throw them with the force of a mass driver. And they make a noisy shatter sound now, too.