Decrease supermatter delam rate in vacuum (#7258)
Added a tweakable multiplier for situations where the supermatter is in no atmosphere (primarily, when someone opens the core vent). Right now, it delaminates in around ten seconds or less from its first announcement (especially if emitter is on) with no chance for engineering to respond. It should take much longer to do so now than before.

In testing, the speed was 4-8% per announcement on a reasonable number of shots (by aurora standards). It will still delaminate, including in space tiles (via core eject, etc, and still assuming it's been shot) but the rate will be less... instant. Turning the emitter on will still naturally increase the rate of damage as it fires.

Also decreased the damage cap some for in-atmosphere damage, also for being able to have time to respond to delaminations (said cap still scales with emitter shots)