Removes the Bicaridine/Kelotane side effects. (#7795)
Why:
The side effects are incredibly annoying right now, knocking you out constantly - somehow being worse than the previous ones. I've thought about replacing them but I can't come up with anything that's not gonna be a pain in the ass for the player, one way or the other. I guess CE_DROPITEM is a possibility but I'm not sure if people would like that.

Thus, I opted to remove them altogether. Slowing down REM is enough in my opinion.