Forensics QOL Changes (#7758)
Makes it so that using the forensics item on certain machinery such as airlocks, alarms, APCs, buttons, windoors etc doesn't apply fingerprints/fibers. This is necessary because these machines will call add_fingerprint BEFORE the forensic kit is used, thus polluting the sample with your own fibers, which is extremely annoying

Also made it so that using a forensics item on a fire alarm or button doesn't make you trigger them, because that's also really annoying.

This is take 2 because the first time I messed up my branches/forks because my brain is smooth.