Adds a new away site with playable ghost roles, Kataphract Chapter (#7744)
Me and Geeves have been working on this for a couple of days now. It's a Kataphract Chapter vessel that can spawn as an away site, with spawners for 2 Kataphract Hopefuls, a Kataphract QM, and a Kataphract Trader. These roles do not require a whitelist, much like the Kosmostrelki. There is also two slots for Kataphract Knights, these require a Unathi whitelist, and a slot for a K'lax Kataphract Hopeful, this requires a Vaurca whitelist.

It has a whole bunch of stuff on it, such as a bridge, sparring/recreation area, QM desk & armoury, vault, knight's office, cafeteria, dorms, toilets, engineering, solars, hangar, and a trading area to trade with the crew and a merchant console, as well as a warehouse to store goods.