Character set up issues for IPC's bugfix (#7926)
As a result of my recent PR. Assigning body types had unforseen changes towards the character set up, namely that bishops had access to hair and that zeng-hu would start with the baseline monitor options.

Changes the way bishop restriction for voidsuit/hardsuit usage.

1- Adds bishops as species_restricted for both voidsuits and hardsuits.
2- Does the same for Zeng-Hu, temporarily the human version.
3- Adjusts bodytype accordingly.

This originally had code for Zeng-Hu usage of machine version of voidsuit/hardsuit, but as I won't be around for a bit, I'll do the bugfix in two phases as I intend to find a less hackey method of fixing this issue. So it changes Zeng-Hu usage of machine version hardsuit/voidsuit for the time being as human version.