Adds sidearms to all command lockers
Gives the CMO, CE and RD the same sidearms that the head of personnel starts out with. Fair play and more smaller weapons for everyone. Goal is to enable the antags easier to grab sidearms (not the armory & bridge ones) while giving the poor command members a fighting chance against the almighty carp.