Mining Tweaks (#8036)
Just a couple small number changes to make mining with pickaxes and drills more viable and decrease the number of credits earned from selling materials. The idea here is to have more send to science and decrease the overall amount of credits spend on guns, or at least require a head of staff to pitch in with their department funds.