Fixes post_droplimb not being called. (#8493)
This makes icons on the body properly update and thus show the missing limbs.

The problem seems to be that the timer doesn't actually spawn at all for whatever reason. Tried 5, 10, nothing works. A spawn(1) worked but I settled for this method as I don't actually think it makes any difference and it works properly.