Adds sec stuff to loadout and removes green trenchcoat. (#8553)
Adds a new security armband and the detective uniforms to loadout. Adjusts the default grey and tan investigator uniforms to have their ties/accents be blue in order to fit in better with the rest of the department. Also fixes the Ftech uniform's pants. This also changes their names to "Tan investigator's uniform" and "grey investigator's uniform" due to both roles being able to wear the outfits. This also removes that green trenchcoat I made because frankly it wasn't very good.