Off brand mad hotfix (#8689)
Just what it says on the tin, small bug, the robots.dmi name just didn't have the right endfix for the module. This is also a test to hopefully clear my commit log so that we don't have to deal with 30 commits on a bugfix/etc. Please be patient if this PR has 30 commits... 👀

This was made in response to this issue. (🎊ALSO WOO ONLY TWO COMMITS 🎊)
Fixes #8685