fixes preferences not accepting flags > 1 << 16 (#8754)
So while I was doing completely different things, I remembered discussing, with Geeves, that our preferences toggles only save 16 bit flags. Despite BYOND now supporting 24 of them. Lo and behold, sanitization is at fault. So I fixed it.

Added a define BITFIELDMAX to indicate the max length of a BYOND bitfield. Replaced all relevant uses of 65535 with it. (Need to do a sweep for 0xffff as well, actually.)