(Optional) Acting Captain Gear (#9198)
    Added an optional acting captain uniform briefcase to the captain's locker. It includes a jacket, a cap, an armband, and a captain radio.

Currently, the only way to show that you're acting captain is to change your ID or to steal the captain's clothes. I thought that was pretty silly, especially for something that's somewhat standard procedure during a chaotic antag round. So I sprited some gear that allows command members to say "Hey, you see this sprite? You should be listening to this guy."