Removes tobacco damage. (#9335)
Fixes #9265

Chewing tobacco causes pain message spam due to the organ damage. This organ damage is on the order of 1e-5 per life tick, and is therefore completely negligible and instantly healed by the player due to the low strength of the tobacco reagent.

In my opinion this damage does not need to be simulated. The tobacco organ damage will never reach a meaningful number in the length of a round even when smoking/chewing tobacco constantly. People overdoing smoking will suffer the nicotine overdose effects.