Dominian gear update/dresses (#9367)
Adds a Dominian mercenary loadout, finishing off the main Human factions that can be represented via mercs.

Also, adds two Dominian dresses to the loadout (the Dominian clothing selection) and fixes the description of the Himeo cap.

The weirdest part of this PR is the 80 melee resists on the Dominian RIG. The other stats have been lowered to compensate for this (given it's meant to be a melee-centered suit) but let me know if the stats aren't to your liking, and I'll be willing to change them.

The Dominian mercenary kit gets 4 voidsuits, the RIG, 4 energy cutlasses and 4 resprited energy shields. The resprite just turns them red.

The solo mercenary kit gives you 1 voidsuit, a shield, a cutlass, and that's it.
The RIG is below, armed and unarmed.