Makes passing arguments to elements work. (#9436)
The components port #8890 only partially implemented a later tg PR #49098 suggested by ninjanomnom here. Missing an important part that affected elements. The result of that was that it wasn't possible to pass any arguments other than the element datum to AddElement().

This PR adds the missing bit, which are the changes in the linked PR here and makes passing arguments to elements work properly.

And another unrelated missing bit that doesn't affect anything (yet)