Adds new synthlore title screens, reworks title logic. (#9407)
Leaving the lobby screen list empty now adds all icon states to the lobby screen list. Also, only one dmi plays at a time.

This means that adding a new title screen is now as simple as just adding it to the dmi. Or making a new dmi for a new category.

Added a new sound when you click things in the menu - more futuristic.