Itching and Saline tweaks.
Saline overdose reduced from 20u to 5u. Coupled with a fast metabolism, this aims to make it so using IV drips is an important step in using Saline, and gives adjusting the transfer rate of IV drips more use instead of sticking them on 4u/sec. and never touching it again.