No more acidic soy sauce.
Replaced the need for sulphuric acid in Soy Sauce with Salt. The kitchen spawns with a moderate amount of salt meaning they can get soy sauce off the bat, however should a chef require more salt, they can bug the pharmacy for some. Salt is also less likely to result in chemists refusing to hand out the chemical, which is the case for sulphuric acid.