More chemical contraband. Medical cabinet content tweak.
Added a new contraband pill bottle - 'Tranquility' - which spawns with neurapan and impedrezene.

Neurapan and Spectrocybin have been added to the list of random reagents which can spawn in vials within maintenance.

The Medical Cabinet by the pharmacy now spawns with 1 bottle of Bicaridine in place of 1 bottle of Dylovene.

The ERT Medical Cabinets now spawn with a bottle of Saline and a pill bottle of Butazoline.

Borers now inject neurapan instead of neurostabin. A stronger anti-psychotic but, if overdosed, can be detrimental to the borer's schemes.