Threshold overdoses, brain regen. changes & feedback changes.
Overdosing now based upon volume in blood, but also requires a small amount of the drug to have metabolised first. Updated chemical ODs & effects to be in-line with this.

Alkysine and Cataleptinol actually heal the brain in different amounts now, instead of both healing ~1% brain activity / tick.
1u of Alkysine - 7.5% brain activity.
1u of Alkysine via drip at a 0.2u/t transfer rate - 10% brain activity.
1u of Cataleptinol -  5%
1u of cryo-cooled Cataleptinol - 10% brain activity.

Reduced the TC cost for the box of sideeffect-be-gone injectors as per feedback.

Changeling Adrenaline Sacs power now injects a small quantity of synaptizine to counter oxycomorphine's side-effects. This'll also make changeling's super strong for ~20 seconds after they use the ability, which will either be a really bad thing or a good thing.

RIG Chemical Injectors now inject 5 units at a time instead of 10 units as per feedback, this will allow more precise dose measurements and to avoid overdosing on synaptizine.

Medibots now inject 10 units of Tricordrazine as opposed to 15 units; chemical bottles now have a default transfer rate of 5 units as opposed to 10 units - changes as per feedback.

Fixed itching as per Matt's review - made a mistake there. Removed the itching emote & damage caused by itching - people can roleplay that if they'd like. The itching messages on top of the emote on top of the pain caused by the itching damage resulted in a fair bit of chatbox cluttering.

Updated chemical effects to be in-line with the new method for overdosing.

Buffed impedrezene a tad as per feedback.

Made butazoline more pink due to feedback saying it was too similar to bicaridine.

Undoes mapping related stuff because that's kind of out of the scope of this pull request.

Lowers mortaphenyl OD to 15u. Lowers oxycomorphine OD to 10u.