Mercs are special bois
As Mercenaries start with the bags from your equipment preferences, this was going to mean people were stuck with the default of backpack. This moves the backpack part of the old equipment prefs, and puts it in the antag part. This pref is now only used if you are a marc.