Makes runes more sneaky and religious.
Per https://forums.aurorastation.org/topic/14730-unapologetic-cult-buff-ideas/ I decided to submit a PR that makes drawing a cult rune done with quill and ink instead of a sacrificial knife you otherwise never see.  The visible text from drawing a rune is subdued, and the selftext is more worshipful.  Runes are not obviously made of blood, but only seem to be; my idea was that the act of charging the blood with arcane energy changes the way it looks, somewhat.