Fix kebab.
Adds rods and brains to the permissible list of items that can be inserted into a cooking container, so you can make the hardcoded kebabs and brain sammiches.

Also fixes a problem with the variable food code that would cause recursion and fail to give the output food a name.