CCIA Recorder Fix
Changes the CCIA Recorder so it can be used to interview other CCIA Agents.
Currently there is a inbuilt protection that prevents a ccia agent from registering himself as the interviewee.
This has been extended to check for the presence of R_CCIA and the absence of a character_id.