Refactors airlock wire messages
Makes it more straightforward to see that all the lights are off (i.e.
no power). Also differentiates between showing bolt status when the
indicators are on vs. off, as one is just a simple visual check while
the other would supposedly require a more close-up inspection.

It also makes it much easier to modify the conditions and text however
needed instead of using a bunch of ternaries for everything.

Additionally index_to_type uses the proper constants instead of
hardcoded numbers.