Adds new mechanics to AI/stationbound control of airlocks - allows AIs to bolt airlocks again, sort of (#10310)
* Refactors airlock wire messages

Makes it more straightforward to see that all the lights are off (i.e.
no power). Also differentiates between showing bolt status when the
indicators are on vs. off, as one is just a simple visual check while
the other would supposedly require a more close-up inspection.

It also makes it much easier to modify the conditions and text however
needed instead of using a bunch of ternaries for everything.

Additionally index_to_type uses the proper constants instead of
hardcoded numbers.

* Lights up the airlock indicator lights

Also does minor refactor of the wires file.

* Adds mechanics to toggle if AIs can bolt doors using AI control wire

* Moves most AI "welcome text" to a separate pop-up window and adds it as a verb

Also adds list of channels to the Radio-Settings verb.

* Moves Initialize, Destroy to beginning, merges the two Initializes

* Adds explicit override for admins and malfs, this allows them to (un)bolt stealthily should they choose to do so

Also adds name and area of the airlock to the UI and refactors it some.

* More better.