NT ER-2 Facelift (#10791)
Revamps the sprites of the energy rifle. Y'know, the one that Odin Security gets. Due to artistic differences with Dream Maker and a desire to finally make this PR after, like, a couple months: Wielded sprites are non-functional. However, the gun doesn't turn invisible when you wield it anymore, so that's good.