adds state_objects
state objects are specifically the objects used to see if the user can interact with the UI. in this case, the appearance_changer datum handles the topic logic and the data, whilst the state_object is set to the specific thing the user clicks on. this allows plastic surgery to work, but not work if you're not adjacent to your patient.