Incubator, Chamber, Capsule, Nullifier
This commit added the UI features for the incubator, as well as adding the Nanomachine Chamber, the main thing where the nanomachines enter the body conventionally. The capsule is just a way of carrying around nanomachine reserves, while the nullifier is a way to cram more programs into your nanomachines.