Better Chamber UI, RnD Capsule Printing
Finally got around to getting rid of the placeholder chamber UI, still missing a couple of things, but if it shipped like this, I wouldn't cry too much. Maybe a little. Also adds the ability for RnD to print more nanomachine capsules, I spoke with a science main I trust to get the material values. We can fine-tune the balance post-merge if we need to.