The Re-bootening
Removed Thigh-high and Knee-high jackboots. They are a pain in the ass to maintain, and were borderline fetish-bait material. Not like you could notice the difference anyway, it was literally one or two pixels, and our spessmen's legs are already short enough for it to not matter.

Anyway. Tweaks the jackboots and oxfords to match the color palettes of the uniforms better. Also, gives the RD, CMO, and OM oxfords too. Because they're command. And they deserve it.