Adds a second rescue RIG to the First Responder's offiice
There are two responders, and only one RIG. The rig is a ton of fun to use, with a slick sprite and a jetpack, leg actuators, etc. However, since there's only one it leads to one FR usually feeling left out especially if the other FR isn't the sharing sort. It isn't fun and it sucks to be in the normal voidsuit while your partner is in the rig,- like driving a model T alongside the batmobile.