Update ore_datum.dm
Increased the "result_amount" of hematite from 5 to 15, coal from 5 to 15, silver from 5 to 10, gold from 5 to 10, , and platinum from 5 to 10.
This is to give better yields for miners since going back and forth from mining to the mining bay is now an up to thirty minute process. Values that aren't multiples of 5 may be more optimal but it's easier for me to think in 5s.
I didn't touch phoron since that's the special funny rock. and I left spread_chance alone because I don't want to make asteroids literally chock full of minerals as far as the eye can see.