Elevator doors see through again.
Basically, reverts the sprite and the opaque change. That's it.

Since, you know. The elevators aren't a death-trap anymore because everyone's forced to be crammed in there.

Because people were deathly allergic to goddamn stairs and ladders on the Old Aurora Map. Like, what the hell? I still don't understand the reasoning behind it. Like, there was already plenty of space and there was already that ladder on maintenance. I mean, unless you didn't have maintenance access. You might as well sit around with a thumb in your ass because either your character can't hack doors or doesn't have access. Better yet, when power was out, you might as well be as good as dead since the elevators didn't work. Even if you're an engineer rolling out from Cryo you'd have to fucking take a hike if the SM was going to delam. Good fucking luck. At least on it's last few months, we finally added goddamn stairs after the old guard rang out. I honestly am amazed at how we were able to operate with such disastorously bad mapping. For years. And like, the key tenets against adding stairs were pretty weak and stupid, and I find it mind-boggling to see that it managed to repel any form of change for so long. What the fuck, man?