Adds Einstein Merc Equipment (Again) (#15799)
* Adds Einstein Merc Equipment

Adds two gear crates for mercenary teams (or an individual) looking to play as Einstein Engines.

The group crate contains two energy rifles, and a heavy bolter. The single crate has one energy rifle. Both come with silenced pistols and themed voidsuits (or a hardsuit in the case of the group crate)

* Minor Syntax Fixes

---------

Co-authored-by: SleepyGemmy <99297919+SleepyGemmy@users.noreply.github.com>