Reallocates the Phoron from the Intrepid (#18818)
- Takes the 30 phoron sheets off of the Intrepid and splits it into 15
for Scientists and 15 for the Machinists.

The Phoron crystals can't actually be used as fuel as they are on the
Intrepid, nor are they able to be used when constructing an RTG as the
circuitboard is not for an advanced RTG. (Only 5 sheets are used in that
recipe anyway, not the 30 in the crate.)

As it stands, I've noticed most Scientists and Machinists split the
Phoron between each other anyhow, since they have no access to it
otherwise, so this just removes that extra step.