Machinist Remap (#18734)
The current machinist bay is extremely tight and hard to maneuver in.
It's shape is odd, and players generally do not like using it.
So, with some creative insight from @NobleRow, I remapped the entire
workshop to replace where RnD and Exploratory chemistry were.

I moved exploratory chemistry and RnD to where the old machinist shop
was. Xenoarch's artifact storage and equipment room has been merged, and
telescience is where the old artifact storage room was.

New Workshop - We do lose some maintenance space, but it is made up for
by the improved playability of an extremely popular role.
I also put the water tank in maintenance because most players do this
already.
![StrongDMM-2024-03-23 14 49
57](https://github.com/Aurorastation/Aurora.3/assets/98699252/b80a45ce-b6a3-4f6f-886d-ba73fc9560a4)



Modified xeno storage - The storage area is entirely independent from
the equipment area. Nothing really changes, aside from layout and
location. I might revisit this and give dedicated lockers instead of the
racks for the suits, however.
![Modified Xenoequipment and
storage](https://github.com/Aurorastation/Aurora.3/assets/98699252/fded1180-c1e5-441d-b0f4-121a7fffeeb2)

New RnD/Chemlab - I tried a different layout to have RnD be visible to
the hall again, but it was very bad due to the 3x3 requirement for the
machinery to optimize QOL. There is less seating in this area, but I
added folding chairs to the science storage area incase they needed more
chairs. This map change does, however, make it so science is more
'consolidated'. RnD is now directly across from the RD's office, and
much closer to the meeting room and storage area.

![ADJUSTEDSCI](https://github.com/Aurorastation/Aurora.3/assets/98699252/e69ff311-6564-4fcc-aa5c-a4d19262809b)


New telescience location - Honestly I just moved it to where it is now
because it is barely played, but would still benefit xeno-arch the most,
I think.

![TelescienceNew](https://github.com/Aurorastation/Aurora.3/assets/98699252/61d1a105-a356-4bd2-9500-749427373a22)