New INDRA Reaction: Iron+Phoron Nucleosynthesis (#18910)
Adds a new INDRA fusion reaction, allowing the production of lead,
uranium and borosilicate glass.

This requires cooperation across departments to receive iron and phoron
from either science or pharmacy.

This is to enable a slow, low yeild manner to obtain uranium, lead and
borosilicate glass in absence of miners, or to recieve a trickle of
uranium, lead, and borosilicate glass that doesn't outpace the payloads
miners can produce.

---------

Signed-off-by: QuestioningMark <49098034+QuestioningMark@users.noreply.github.com>