The Xanu Frigate (#18516)
WIP.

Adds the All-Xanu Spacefleet Shrike-class frigate, a comparatively
colossal warship in service with Xanu Prime's navy. Built for combat, it
has a myriad of features designed to make it stand toe-to-toe against
most enemy combatants.

It features:

- Separate navigation bridge and CIC, a la the Horizon's command bunker.
The CIC has its own air and power supply.
- Redundant thruster nacelles; even if one nacelle is destroyed, the
ship can continue operating on a single one.
- Two fusion reactors. The power required by this thing is just that
huge.
- Thiccccc armor plating and whipple shields made of grating to protect
against meteor and dust strikes.
- A Longbow and Grauwolf.
- A fightercraft shuttle, with a Francisca.
- A boarding shuttle.
- Fully-stocked medbay with operating room.
- A brig.

The ghostrole capacity for this is 7 (6 crew + 1 captain), but has
capacity for all the way up to 15 (14 crew + 1 captain) during events.

The intended purpose of this frigate is twofold: a reduced-power
"patrol" configuration for light skirmishes, and a full-armament
"battle" configuration so that the frigate can act as an offship ERT in
case the Horizon gets into trouble, either against another party or
against the Coalition. If used in its battle configuration it can
support 15 crew at full capacity.
<details>
<summary>Screenshots</summary>


![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/8c8ade90-947e-49c8-b710-0997fcf1cb0f)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/c105e515-abf4-42b1-8f8d-d8de5d6692fc)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/e165eb14-6976-472a-b6ab-e243eca9d4ec)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/b811e586-cb75-467b-ba38-2c737221d69e)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/b198b50c-7753-4aaa-839e-eb57a6bcedec)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/00029766-664c-41d0-b075-05cdca325dff)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/bf552c19-7c3d-4d5d-b713-916e3e0b2de3)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/3d92a57d-bb72-4c53-86de-fe218c5fe439)

</details>