Сonstructable with rods grills (#18751)
I was once asked to install a grille in a window. However, I was not
able to do it because the code does not provide for it. I had the face
of Pepe the frog.

Refactors the code a bit and adds the ability to use rods on windows to
create grids. You will need two rods for installation. Remember that the
glass will interfere with the installation.

---------

Signed-off-by: SidVeld <sidveld@gmail.com>
Co-authored-by: SleepyGemmy <99297919+SleepyGemmy@users.noreply.github.com>