Uueoa-Esa, Part 2: System Ruins (#18944)
Adds several ruins for the uninhabited planets of Uueoa-Esa.
Adds a pre-contact Unathi spacesuit for one of the ruins and potential
future use.
Voidsuits and other refittable clothing items (if any are added) can now
be set to refit to a species on initialize rather than copypasting code
for mapped-in suit variants.

Will require updating for #18850 and #18920 when merged.