Translator augment (#18905)
Adds an augmented translator to the game, with which the user can
understand all human languages. I did not add a translator from xeno
languages, as it would have been very strong, given the difficulty of
translation and the peculiarities of their languages.

I have decomposed the original augment file to improve readability and
code handling. Hopefully this won't cause trouble in the future. (Sorry,
but an 800+ line file is a very large file.)