Luceism Additions (#18906)
Adds a lot of Luceian stuff for Assunzionis and brings them up to speed
as a contender religion.

- Added several items relating to Luceism, the religion of Assunzione.
- Added chaplain items for Luceism: robes, a cassock, and a new null
rod, the Luminous Sceptre.
- Added a new versebook, the Luceian Book of Scripture, available to
Assunzionii characters. It's a versebook with various texts from
Luceism.
- Resprited the Luceian amulet, Luceian Book of Scripture, and the
Assunzione cloak.
- Compressed the loadout warding sphere and warding sphere case into
just one loadout item that already comes with a case.
- Religions with unique religious books (versebooks) now have their
chaplains spawn with the versebook in question, rather than a generic
black tome.


![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/e477d9c7-cbfb-4976-a905-1bda0cd99820)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/7523e3c8-04ee-4146-ae9f-93702020adda)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/68a69aa5-a346-4c38-b891-73516a877615)