Adds the Scarab Salvager to the Badlands and Valley Hale (#19001)
This adds the Scarab Salvager offship to the Badlands and Valley Hale
sectors. Has assent from human lore, see
[here.](https://discord.com/channels/724651070017765459/724653076866400336/1229465269903036457)
They're space nomads so I feel it's reasonable for their range to be a
little expanded to non-developed sectors.

Additionally, fixed a few mapping mistakes in the Hivebot Hub and Scarab
Salvager maps.