Pirate Base Tweaks (#18849)
Gives the pirate base pirate shuttle a proper docking airlock rather
than the deprecated spawner.
Replaces the syndicate spacesuits on the pirate base with a selection of
random ones.

Changes the pirate base description to not immediately tell anyone
scanning it that there are pirates there so that the Horizon can doesn't
have a massive IC reason to avoid it.
Gives the pirate base areas actual names, because for some reason they
didn't have them.

---------

Signed-off-by: Matt Atlas <mattiathebest2000@hotmail.it>
Co-authored-by: Matt Atlas <mattiathebest2000@hotmail.it>