Bugfix: Fixes the power for the freebooter ship (#19066)
I was told by some players that the Freebooter ship power did not work
and required players to rewire the powernet. This was because the wrong
wire was mapped in. A "power knot" was wrongfully used, it had a
direction set instead of the proper icon state. It's been fixed, tested
and now works.