Canary minor qol (#19011)
Flips the sensor and landing control sprites for the canary consoles so
they'd fit on the opposite sides, swaps the pipes around for the fuel to
thrusters, to accomodate moving the pump for the fuel can inside, and
shifts the sensors and engine control consoles around so the WSO has the
Sensors and gun, and the pilot has the engines, landing control, and
helm.

Requested by Firstact.