Consolidates text files into their own, central folder (#19063)
Sorts the various .TXTs of the game (currently so far the lore radio
stations as well as the "verse books" (Biesel constitution, religious
scriptures, etc)) to a master `texts` file in Aurora's root directory.
Should help with config issues as well as consolidate all the texts in
one place, especially for use in future endeavors where bits of text
like these need to be made.

Should theoretically fix #19060. If it doesn't then, well, I'm stumped.