Sensor relay refit (#19021)
Overhauls the sensor relay away site to be up to modern standards. It's
now more practical, more decrepit, and has more facilities for things
while retaining the old loot and goodies it had before.

The place starts depowered, save for the airlocks, which are powered by
default to allow entry without having to bust down a
window/wall/inflatable and vent the place.

In addition this PR adds a special sprite for the sensor array the relay
uses, and tweaks its stats:
* Now has a maximum range of 24.
* Has a stable range of 10.
* Cannot deep scan at all.
* Overheats very quickly at high ranges.
* Detects passing ships slower than a standard relay.

The above should make it more worthwhile to capture and use, as when
datalinked it provides visibility of a large portion of the sector.


![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/e53fd13e-ca7c-402e-9605-c213f729df7d)


![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/8d09db8e-b688-4dca-84b1-2945dc70f702)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/9e644f4c-1cf4-4fbc-8119-1cbfbba59a15)

# Before powering

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/156b61bd-eb7b-4cfc-8eca-5a752c4c4bea)

# After powering
<details><summary>Click to expand</summary>
<p>


![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/357f01fc-c0ce-46ca-9869-3ec4bdd6a14e)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/b3f74c29-14ad-4765-94b8-81b0591e5916)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/d155dd5f-69ff-4171-a0e0-ac871cff3415)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/e1db7407-8e7a-4c54-bfe4-0103000fc317)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/ffb63725-a34f-4743-a82c-c780a11421c0)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/59aeedae-7d3c-458b-bea5-5fe8cc94ebf0)

![image](https://github.com/Aurorastation/Aurora.3/assets/55491249/10f2cfba-fb80-4f2e-b8e4-d74a95f6830c)

</p>
</details>