Wiring Guidelines map (#18982)
This map is to provide the proper wiring palette, how to setup basic
power nets - with examples of what not to do as well.

I include two miniature stations, one with proper wiring, and the second
with improper wiring - per our mapping guidelines.

Additionally I added several examples of what to do, and what not to do
when routing wires into buildings, as well as setting up electrified
windows.