[Bugfix] Minor additions and turf fix for the AI core (#19079)
I did not realize that there was a missing floor tile from the scuttle
chamber. It was covered by decals in SDMM and I did not notice it during
the test-merge.

I also overlooked adding a holopad to the AI upload. 

Additional changes: Three extra cameras that I had intended to add, but
it slipped my mind. Sorry!